Fixes issue 22 - sanitize mis-formats numbers
The sanitize function converts numbers (actually, all non-strings,
non-lists) to a string (using to-s) before they get to the format
control string, and to-s uses the ~a directive. This means the control
string provided to format-wish cannot use numeric directives.

This fix adds a method to handle numbers by passing them back unchanged
for formatting by the format-wish control string.