- ensured validation mode that causes trouble are not used
There is a problem when using :validate :all or ':validate focus' in a
entry or a spinbox.

If the next action after the entry loosed focus is a click on a menu, and then on
a menu button, the program freezes.

This behaviour will not happen if validate is :focusout, though.