- fixed 'stealing' of stop-thread flag.
for future reference:

The demo  has a structure  that contains  the rendering thread  and a
flag  that, in  not null,  stop the  rendering loop  contained in  the
aforementioned thread

```
(defstruct animation
  (thread)
  (lock (make-lock))
  (stop-p nil))
```

the thread contains code similar to:

```
(loop while (not (animation-thread *animation*)) do
      ;; draw stuff
)
```

where *animation* is  a special variable that is bound  to the current
animation (in this  demo only a single animation can  be active at the
same time)

when a radio button of the demo UI has been clicked, the old animation
is stopped and a new is started

```
(stop-animation)
```

the code above will set the slot stop-p  of the object *animation* to
't', then will wait the rendering loop to exit and the rendering
thread to join, finally it returns

```
(setf *animation*
      (make-animation :thread (make-thread #'draw-plasma-thread
                                           :name "plasma"))))
```

But this code has a subtle bug: as we are creating a thread inside the
call to create a struct, will be first evaluated the code to start the
rendering thread or the code to create the struct?

If the  second occurs  no problem  arises but, if  the former  is true
(thread first, struct after) the new  thread will start before the new
struct is created so the rendering loop will look at the slot 'stop-p'
of the  old value of *animation*  that is, in fact  (because we called
the function 'stop-animation' before), non null. So the rendering loop
immediately  exits and  so the  rendering thread;  the results  is the
animation does not even starts

My solution  is to separate the  creation the struct (that  occurs for
first) and then start the rendering thread:

```
(setf *animation* (make-animation))
(setf (animation-thread *animation*)
      (make-thread #'draw-plasma-thread
                   :name "plasma"))
```